jv - remove old userlist components
