+++ client/src/components/profiles/ProfilePage.tsx
@@ -4,9 +4,8 @@ import { EditIcon } from "@chakra-ui/icons";
 import { Box, Flex, Heading, HStack, Link, VStack } from "@chakra-ui/layout";
 import { Avatar, Button, ButtonGroup, IconButton, Tooltip } from "@chakra-ui/react";
 import { useToast } from "@chakra-ui/toast";
-import { data } from "msw/lib/types/context";
 import React from "react";
-import { Block, BlockInput, EmbeddedUserList, MalLinkOauthDocument, MalLinkOauthQuery, ProfileDocument, User, useUpdateProfilePageBlocksMutation } from "../../generated/graphql";
+import { Block, BlockInput, EmbeddedUserList, MalLinkOauthDocument, MalLinkOauthQuery, ProfileDocument, useUpdateProfilePageBlocksMutation } from "../../generated/graphql";
 import { convertBlockToBlockInput } from "../../utils/convertBlock";
 import useImperativeQuery from "../../utils/useImperativeQuery";
 import AddBlockModal from "./AddBlockModal";
